feat: Add change canvas color feature
Added function to change the color of the canvas without reset
Added several helper function to convert color strings
Added a helper function to detect if color is light.
Color and Change Canvas button make use of that funtion.
Switched to using an event listener on a button to trigger the
color picker.